Cost functions take floats not ints
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34076 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ bindings/java-jni/jni-client.h
@@ -23,7 +23,7 @@ JNIEXPORT void JNICALL Java_org_linkgrammar_LinkGrammar_setMaxParseSeconds
  * Signature: (I)V
  */
 JNIEXPORT void JNICALL Java_org_linkgrammar_LinkGrammar_setMaxCost
-	(JNIEnv *, jclass, jint);
+	(JNIEnv *, jclass, jdouble);
 
 /*
  * Class:     LinkGrammar
